Enable syslog by default
This gives more relevant data to journal than logging in stdout. Also, pid file
is no longer needed, since we're a Type=dbus service.

The flat logging file is removed since it's not needed when we're logging into
journal. We probably can remove the logrotate file too.

The -u argument is moved into the service file as we can't really remove it
without changing the service's Type.